finally organized compnents to the point where we do not fail to build
